+++ app/src/main/java/de/passwordvault/viewmodel/activities/SecurityQuestionViewModel.java
@@ -1,5 +1,7 @@
 package de.passwordvault.viewmodel.activities;
 
+import android.util.Log;
+
 import androidx.lifecycle.ViewModel;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -52,6 +54,34 @@ public class SecurityQuestionViewModel extends ViewModel {
             questionStrings.add(allQuestionStrings[q.getQuestion()]);
         }
         return questionStrings;
+    }
+
+
+    /**
+     * Method tests whether the provided answers are correct. The index of each answer within the
+     * passed list must correspond to the index of it's respective question within the list returned
+     * by {@link #getQuestions()}.
+     *
+     * @param answers   List of answers to test.
+     * @return          Whether the answers are correct.
+     */
+    public boolean areAnswersValid(ArrayList<String> answers) {
+        if (answers.size() != questions.size()) {
+            return false;
+        }
+        Log.d("SQ", "Number of answers=" + answers.size() + ", number of questions=" + questions.size());
+        int numberOfCorrectAnswers = 0;
+        for (int i = 0; i < answers.size(); i++) {
+            if (answers.get(i) == null) {
+                Log.d("SQ", "null");
+                continue;
+            }
+            Log.d("SQ", "Testing answer '" + answers.get(i) + " with correct value '" + questions.get(i).getAnswer() + "'.");
+            if (answers.get(i).equalsIgnoreCase(questions.get(i).getAnswer())) {
+                numberOfCorrectAnswers++;
+            }
+        }
+        return numberOfCorrectAnswers >= Account.REQUIRED_CORRECT_ANSWERS;
     }
 
 }